Moved lib code to pkg/verso
