+++ pkg/verso/extract.go
