+++ pkg/verso/extract_test.go
