+++ pkg/verso/parse.go
